+++ src/models/group/group.ts
@@ -1,0 +1,15 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { NumericID } from '../../value-objects';
+
+export interface Group {
+	id: NumericID;
+	name: string;
+	description: string | null;
+}
