+++ src/models/group/is-group.ts
@@ -7,26 +7,8 @@
  **************************************************************************/
 
 import { isNull, isString } from 'lodash';
-import { NumericID, isNumericID } from '../value-objects';
-
-export interface RawGroup {
-	GID: number;
-	Name: string;
-	Desc: string; // Empty is null
-	Synced: boolean;
-}
-
-export interface Group {
-	id: NumericID;
-	name: string;
-	description: string | null;
-}
-
-export const toGroup = (raw: RawGroup): Group => ({
-	id: raw.GID.toString(),
-	name: raw.Name,
-	description: raw.Desc.trim() === '' ? null : raw.Desc.trim(),
-});
+import { isNumericID } from '../../value-objects';
+import { Group } from './group';
 
 export const isGroup = (value: any): value is Group => {
 	try {
