+++ src/models/group/raw-group.ts
@@ -1,0 +1,14 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface RawGroup {
+	GID: number;
+	Name: string;
+	Desc: string; // Empty is null
+	Synced: boolean;
+}
